+++ CHANGES.md
@@ -3,6 +3,7 @@
 **0.5.4**
 
 * StreamServer.run() is deprecated. Use start(), startSecure() or startOn() instead.
+* URL mapping supports grouping, such as /user/(name:[^/]*)
 * Support the new Dart I/O.
 
 **0.5.3**